+++ back/src/main/java/com/curso/spring/rest/models/dao/FacturaDao.java
@@ -12,10 +12,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FacturaDao extends JpaRepository<Factura, Long> {
+public interface FacturaDao extends JpaRepository<Factura, Integer> {
 
     @Query(value = "select f from Factura f join fetch f.cliente c join fetch f.items i join fetch i.producto where f.id = ?1")
-    Factura fetchByIdWithClienteWithItemFacturaWithProducto(Long id);
+    Factura fetchByIdWithClienteWithItemFacturaWithProducto(Integer id);
 
     Page<Factura> findAllFacturasByClienteOrderById(Cliente cliente, Pageable pageable);
